backend_subscription/requirements.txt
django
djangorestframework
django-filters
